Create initial html pages for Odin assignment
This is the initial commit for the Odin Project recipe pages assignment.
This code establishes an index.html page that links to indivual pages
for a group of recipes. Each recipe page contains an image of the completed
dish, a short description, necessary ingredients, and the directions for making
the recipe. A link back to the index page was also added for ease of use.